consulta por ID

+++ Curso PHP/Oficina/painel-adm/fornecedores/excluir.php
@@ -3,9 +3,16 @@ require_once("../../conexao.php");
 
 $id = $_POST['id'];
 
+//EXCLUIR SOMENTE SE NÃO TIVER REGISTROS RELACIONADOS
+$query_tot = $pdo->query("SELECT * FROM produtos where fornecedor = '$id' ");
+$res_tot = $query_tot->fetchAll(PDO::FETCH_ASSOC);
+$total_produtos = @count($res_tot);
 
-$pdo->query("DELETE FROM fornecedores WHERE id = '$id'");
-
-echo 'Excluído com Sucesso!';
+if($total_produtos == 0){
+	$pdo->query("DELETE FROM fornecedores WHERE id = '$id' ");
+	echo 'Excluído com Sucesso!';
+}else{
+	echo 'Você possui produtos relacionados a este fornecedor, primeiro exclua os produtos para depois excluir a fornecedor!';
+}
 
 ?>
